@@ -12,7 +12,7 @@
     "react": "15.3.2",
     "react-dom": "15.3.2",
     "react-redux": "4.4.5",
-    "react-select2-wrapper": "^1.0.4-beta5",
+    "react-select2": "4.0.3",
     "redux": "3.6.0",
     "redux-logger": "2.7.4",
     "redux-promise": "0.5.3",
@@ -21,12 +21,12 @@
   },
   "license": "MIT",
   "engines": {
-    "node": "5.12.0"
+    "node": "6.11.4"
   },
   "devDependencies": {
     "es6-object-assign": "1.0.3",
     "react-addons-test-utils": "15.3.2",
     "sinon": "1.17.7",
-    "webpack-dev-server": "^2.9.1"
+    "webpack-dev-server": "2.9.1"
   }
 }
