@@ -1,6 +1,5 @@
-const DateFilter = require('../helpers/date_filters')
+import DateFilter from '../helpers/date_filters'
 
-const calendarDF = new DateFilter("calendar_filter_btn", "Tous les champs du filtre de date doivent être remplis", "#q_contains_date_NUMi")
+const calendarDF = new DateFilter("calendar_filter_btn", "Tous les champs du filtre de date doivent être remplis", "q_contains_date_NUMi")
 
-module.exports = calendarDF
-
+export default calendarDF
