@@ -1,5 +1,5 @@
-const DateFilter = require('../helpers/date_filters')
+import DateFilter from '../helpers/date_filters'
 
-const complianceControlSetDF = new DateFilter("compliance_control_set_filter_btn", "Tous les champs du filtre de date doivent être remplis", "#q_updated_at_start_date_NUMi", "#q_updated_at_end_date_NUMi")
+const complianceControlSetDF = new DateFilter("compliance_control_set_filter_btn", "Tous les champs du filtre de date doivent être remplis", "q_updated_at_start_date_NUMi", "q_updated_at_end_date_NUMi")
 
-module.exports = complianceControlSetDF
+export default complianceControlSetDF
