@@ -1,5 +1,5 @@
-const DateFilter = require('../helpers/date_filters')
+import DateFilter from '../helpers/date_filters'
 
-const timetableDF = new DateFilter("time_table_filter_btn", "Tous les champs du filtre de date doivent être remplis", "#q_bounding_dates_start_date_NUMi", "#q_bounding_dates_end_date_NUMi")
+const timetableDF = new DateFilter("time_table_filter_btn", "Tous les champs du filtre de date doivent être remplis", "q_bounding_dates_start_date_NUMi", "q_bounding_dates_end_date_NUMi")
 
-module.exports = timetableDF
+export default timetableDF
