@@ -7,4 +7,5 @@
 // To reference this file, add <%= javascript_pack_tag 'application' %> to the appropriate
 // layout file, like app/views/layouts/application.html.erb
 
-console.log('Hello World from Webpacker')
+// import jQuery from 'jquery'
+// import {} from 'jquery-ujs'
